requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
aiohttp==3.4.4
async-timeout==3.0.0; python_version >= '3.5.3'
attrs==18.2.0
chardet==3.0.4
click==6.7
dnspython==1.15.0
flask==1.0.2
gprof2dot==2017.9.19
idna-ssl==1.1.0; python_version < '3.7'
idna==2.7
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
multidict==4.4.2; python_version >= '3.4.1'
numpy==1.15.2
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
werkzeug==0.14.1
yarl==1.2.6; python_version >= '3.4.1'
